Refactor help state and improve runtime performance
